requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-gis==0.18
docopt==0.6.2
GDAL @ file:///C:/Users/DELL/Downloads/GDAL-3.4.3-cp38-cp38-win_amd64.whl
idna==3.3
Js2Py==0.71
leaflet==0.0.3
munch==2.5.0
numpy==1.22.3
packaging==21.3
pandas @ file:///C:/Users/DELL/pipwin/pandas-1.4.2-cp38-cp38-win_amd64.whl
Pillow==9.1.0
pipwin==0.5.2
psycopg2-binary==2.9.3
pyjsparser==2.7.1
pyparsing==3.0.8
PyPrind==2.11.3
pyproj==3.3.1
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
Rtree==1.0.0
Shapely @ file:///C:/Users/DELL/pipwin/Shapely-1.8.1.post1-cp38-cp38-win_amd64.whl
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
